Refactor package structure and improve code logic
