+++ url_snapshotter/__init__.py
@@ -1,1 +1,1 @@
-# This file makes the directory a package.
+# url_snapshotter/__init__.py
